requirements.txt
python-telegram-bot>=20.0
hypothesis
pytest
pytest-asyncio
python-dotenv
